requirements.txt
colorama==0.4.6
cryptography==45.0.4
python_bcrypt==0.3.2
